Added a message to the confirmations list when your session expires

+++ Steam Desktop Authenticator/BrowserRequestHandler.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Windows.Forms;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -45,7 +46,14 @@ namespace Steam_Desktop_Authenticator
         }
 
         CefReturnValue IRequestHandler.OnBeforeResourceLoad(IWebBrowser browserControl, IBrowser browser, IFrame frame, IRequest request, IRequestCallback callback)
-        {
+        {   
+            // Check if the session is expired
+            if (request.Url == "steammobile://lostauth")
+            {
+                MessageBox.Show("Failed to load confirmations.\nTry using \"Force session refresh\" under the Selected Account menu.\nIf that doesn't work use the \"Login again\" option.", "Confirmations", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                return CefReturnValue.Cancel;
+            }
+            
             // For some reason, in order to set cookies manually using a hdeader you need to clear the real cookies every time :/
             Cef.GetGlobalCookieManager().VisitAllCookies(new DeleteAllCookiesVisitor());
